% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anolis_map.R
\docType{data}
\encoding{UTF-8}
\name{Anolis_map}
\alias{Anolis_map}
\title{Anolis map of regimes}
\format{A two columns matrix giving the evolutionary time spent by each branch of \code{\link{Anolis_tree}} in the two islands
C: Cuba and H: Hispniola}
\usage{
data(Anolis_map)
}
\description{
Anolis map of regimes
}
\details{
This data contains a map of the Anolis lizards' clades geographic position onto the Anolis phylogeny.
A two columns matrix giving the evolutionary time spent by each branch of \code{\link{Anolis_tree}} in the two islands
C: Cuba and H: Hispniola. Each row represents a branch of \code{\link{Anolis_tree}}, numbers represent the evolutionary time spent on each island by the clade.
}
\author{
Theo Gaboriau
}
\keyword{data}
