% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anolis_traits.R
\docType{data}
\encoding{UTF-8}
\name{Anolis_traits}
\alias{Anolis_traits}
\title{Anolis snout to vent length}
\format{A three columns data.frame with species names (species) and snout to vent length (svl) and elevation of individual observations}
\usage{
data(Anolis_traits)
}
\description{
Anolis snout to vent length
}
\details{
This data has been extracted from the study of Munoz et al., 2014
It contains individual observations of Anolis lizards' snout to vent length (mm) and elevation (m) from Hispaniola and Cuba islands in a three columns data.frame with species names (species) and snout to vent length (svl) and elevation of individual observations.
}
\references{
Munoz, M. M., Wegener, J. E., and Algar, A. C. (2014).
 Untangling intra- and interspecific effects on body size clines reveals divergent processes structuring convergent patterns in Anolis lizards.
  Am. Nat. , 184(5):636-46.
  
  Munoz, Martha M., Wegener, Johanna E., Algar, Adam C. (2014),
  Data from: Untangling intra- and interspecific effects on body size clines reveals divergent processes structuring convergent patterns in Anolis lizards,
  Dryad, Dataset, https://doi.org/10.5061/dryad.q39h2
}
\author{
Martha Munoz, Johanna Wegener and Adam Algar.
}
\keyword{data}
