% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anolis_tree.R
\docType{data}
\encoding{UTF-8}
\name{Anolis_tree}
\alias{Anolis_tree}
\title{Anolis phylogenetic tree}
\format{A dated phylogenetic tree of class "phylo" including 16 species of Anolis lizards}
\usage{
data(Anolis_tree)
}
\description{
Anolis phylogenetic tree
}
\details{
This phylogenetic tree as been extracted from the Anolis phylogeny of Poe et al., 2017
It contains dated phylogenetic tree of class "phylo" including 16 species of Anolis lizards
}
\references{
Poe, S., Nieto&#45;Montes de Oca, A., Torres&#45;carvajal, O., De Queiroz, K., Velasco, J. A.,
 Truett, B., Gray, L. N., Ryan, M. J., Kohler, G., Ayala&#45;varela, F., and Latella, I. (2017).
  A Phylogenetic, biogeographic, and taxonomic study of all extant species of Anolis (Squamata, Iguanidae).
  Syst. Biol., 66(5):663-697
}
\author{
Theo Gaboriau
}
\keyword{data}
