% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_logml.R
\name{comp_logml}
\alias{comp_logml}
\title{Compute Maximum log-Likelihood}
\usage{
comp_logml(ySy, XSy, invBN, indic, iB0)
}
\arguments{
\item{ySy}{matrix product of y vector normalized by standard deviation}

\item{XSy}{matrix product of data, normalizing matrix and y outcome}

\item{invBN}{inverse Beta matrix}

\item{indic}{indices of likelihood to be accessed}

\item{iB0}{previous precision matrix}
}
\value{
Returns an object containing the numeric value of the marginal Likelihood and beta parameters
}
\description{
[Internal Function] \code{comp_logml} will return a numeric value, the maximum log-likelihood
y = X*alpha+epsilon,  epsilon N(0,S), alpha ~ N(0,inv(iA0))
}
\details{
ySy ...  y'S^{-1}y
XSy ...  X'S^{-1}y
invBN=X'S^{-1}X+ iB0
iB0     k x  k prior precision matrix
indic    k indicators for regression effects
}
\seealso{
\code{\link{lcondx}}
}
