% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_trt_effects.R
\name{comp_trt_effects}
\alias{comp_trt_effects}
\title{Compute treatment effects}
\usage{
comp_trt_effects(mcmc_select, thin = 1)
}
\description{
[Internal Function] computes treatment effects from the coefficients
sampled in the MCMC process.
}
