% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condmoms_util.R
\name{condmoms_util}
\alias{condmoms_util}
\title{Conditional Moments Computation}
\usage{
condmoms_util(mu_xst, resy, start_x, nx, start_y, sgma, rho,
  lambda = NULL)
}
\arguments{
\item{mu_xst}{mean vector of latent utility values}

\item{resy}{residual vector on y}

\item{start_x}{matrix of Tmax x trtmax holding information at which indicator
value observations with panel time t and treatment j start}

\item{nx}{T x trtmax matrix containing number of observations for each combination of t and j}

\item{start_y}{analoguous to start_x for outcome values}

\item{sgma}{variance matrix}

\item{rho}{covariance matrix}

\item{lambda}{(optional) reserved parameter for further programming}
}
\value{
Returns a list of moments of centered x:
        \enumerate{
        \item conditional mean
        \item conditional variance
}
}
\description{
Internal. \code{condmoms_util} returns an object consisting of conditional moments
}
\details{
Computes conditional mean and variance in the utility x* model.
Construct array resyt (Residuals of y_t):
n x Tmax with zeros for not observed values
}
