% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_errorvar.R
\name{draw_errorvar}
\alias{draw_errorvar}
\title{Draw Error Variance}
\usage{
draw_errorvar(eps2, indt, indy0, sn, prior)
}
\arguments{
\item{eps2}{vector of squared residuals}

\item{indt}{indicators where panel times start.}

\item{indy0}{index y vector}

\item{sn}{previous matrix to determine correct dimension of variance matrix}

\item{prior}{object containing prior information, prior standard deviance is used}
}
\description{
[Internal Function] \code{draw_errorvar} samples the variance of the errorterm from a (conjugate)
}
