% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_factor_switchreg.R
\name{draw_factor_switchreg}
\alias{draw_factor_switchreg}
\title{Draw factors for SRF/SRI}
\usage{
draw_factor_switchreg(resy, resx, start_x, nx, start_y, sgma, rho, lambda,
  D)
}
\description{
Draw factors for SRF/SRI
}
