% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_indic_beta_lambda.R
\name{draw_indic_beta_lambda}
\alias{draw_indic_beta_lambda}
\title{Draw BETA Indicators and Coefficients}
\usage{
draw_indic_beta_lambda(y, X, sgma2, delta, deltafix, omega1, omega2, dy, invB0,
  isel)
}
\arguments{
\item{y}{outcome vector}

\item{X}{data matrix}

\item{sgma2}{variance}

\item{delta}{vector of indices which model parameters are exposed to
the spike and slab variable selection process}

\item{deltafix}{vector indicating which variables are fixed}

\item{omega1}{omega parameter of beta}

\item{omega2}{omega parameter of lambda}

\item{dy}{dimension of outcome}

\item{invB0}{startvalue matrix of beta coefficients}

\item{isel}{indicator whether selection is performed}
}
\description{
\code{draw_indic_beta_lambda} draws beta coefficients for the outcome
models (y) and lambda parameters for shared factor modelling.
}
\details{
y....Tn x1 ...response from fixed effects model
x .... nx 1 ... treatments
X ....Tn x df regressor matrix
}
