% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_indic_coeff_switchreg.R
\name{draw_indic_coeff_switchreg}
\alias{draw_indic_coeff_switchreg}
\title{Draw indices for selection upon sigma and rho parameters for SRF/SRI models}
\usage{
draw_indic_coeff_switchreg(x, y, Wx, Wy, start_x, nx, start_y, sgma, rho,
  lambda, D, delta, deltafix, omega1, omega2, pri_invB0, isel)
}
\description{
Internal.
}
