% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_loadings.R
\name{draw_loadings}
\alias{draw_loadings}
\title{Draw Factor Loadings}
\usage{
draw_loadings(resy, Xf, resx, start_x, nx, start_y, sgma, rho, prior)
}
\arguments{
\item{resy}{residual vector on y outcome}

\item{Xf}{factor matrix of features and latent factor f}

\item{resx}{residual vector on x}

\item{start_x}{matrix containing indices where certain observations
begin in the vector sorted by treatment and panel times}

\item{nx}{number of x observations}

\item{start_y}{matrix containing starting indices of outcomes}

\item{sgma}{standard deviation}

\item{rho}{correlation coefficient}

\item{prior}{list object containing prior information}
}
\value{
lambda ... n factor loadings
}
\description{
Function for drawing factor loadings random effects
}
