% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_indic_alphaSF.R
\name{drawAlphaIndicesSF}
\alias{drawAlphaIndicesSF}
\title{Draw indicators and regression coefficients}
\usage{
drawAlphaIndicesSF(y, X, delta, deltafix, omega, invA0, isel,
  fix.alpha = FALSE)
}
\arguments{
\item{y}{outcome vector}

\item{X}{data matrix}

\item{delta}{delta index of selection}

\item{deltafix}{fixed delta indices, not subject to variable selection}

\item{omega}{correlation parameter}

\item{invA0}{inverse alpha entries matrix}

\item{isel}{indicator whether selection is performed}
}
\value{
delta ...  k x 1 indicators

alpha ... k x 1  regression effects
}
\description{
[Internal Function] \code{draw_indic_alphaSF} draws indicator values and
regression coefficients for the selection model (probit model of
choosing treatment)
}
\details{
The procedure also includes variable selection process, utilizing a
Dirac spike prior distribution for variable selection
single move update
y... n x 1 response from fixed effects model
X... n x k regressor matrix
}
