% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawErrorVariance.R
\name{drawErrorVariance}
\alias{drawErrorVariance}
\title{Draw Error Variance}
\usage{
drawErrorVariance(eps2, indt, indy0, sn, prior, fix.sigma = FALSE)
}
\arguments{
\item{eps2}{vector of squared residuals}

\item{indt}{indicators where panel times start.}

\item{indy0}{index y vector}

\item{sn}{previous matrix to determine correct dimension of variance matrix}

\item{prior}{object containing prior information, prior standard deviance is used}
}
\description{
Internal.
}
