% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sharedfactor.R
\name{drawSharedFactor}
\alias{drawSharedFactor}
\title{Draw Shared Factor}
\usage{
drawSharedFactor(epsx, lambdax, epsy, sgma2, lambda, start_x, nx, start_y,
  Tn, fix.f = FALSE)
}
\arguments{
\item{epsx}{residual vector on x}

\item{lambdax}{lambda parameter of selection model x*}

\item{epsy}{residuals on y}

\item{sgma2}{variance parameters}

\item{lambda}{lambda parameters on outcome for each panel time and treatment}

\item{start_x}{indicator matrix}

\item{nx}{number of observations}

\item{start_y}{indicator matrix on y}

\item{Tn}{number of observations times panel t}
}
\value{
f ... shared latent factor
}
\description{
[Internal Function] \code{draw_sharedfactor} draws the latent factors \eqn{f_i} for
SF model.
}
