% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_utility.R
\name{drawUtility}
\alias{drawUtility}
\title{Draw latent utilities}
\usage{
drawUtility(x, mxst, sxst)
}
\arguments{
\item{x}{vector of subject features}

\item{mxst}{vector of means of utilities x*}

\item{sxst}{vector of standard deviations of utilities x*}
}
\value{
xst x*, the latent utility for selection model
}
\description{
[Internal Function] \code{draw_utility} draws latent utilities for the
probit model, which represents the selection of treatment. Latent
utilities (x*) range from -Inf to +Inf. For x* being positive, the
subject selects treatment, subject does not decide for treatment otherwise.
}
