% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{ineff_factor}
\alias{ineff_factor}
\title{Inefficiency Factor}
\usage{
ineff_factor(x)
}
\arguments{
\item{x}{a time series of length k}
}
\value{
2 numeric values, the empirical inefficiency factor and m
}
\description{
\code{ineff_factor} calculates inefficiency facors for certain time series
with autocorrelation after Geyer (1992).
}
