% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jiveMake}
\alias{jiveMake}
\title{Make jive object}
\usage{
jiveMake(simmap, traits, model.var = "OU1", model.mean = "BM",
  model.lik = "Multinorm", map = NULL, root.station=TRUE, 
  scaleHeight=FALSE)
}
\arguments{
\item{simmap}{an object of class "jive" (see details)}

\item{traits}{name of the output file that will store the log of MCMC chain}

\item{model.var}{sampling frequency of the MCMC chain (how often chain will be saved into output file}

\item{model.mean}{printing frequency of the MCMC chain (how often chain will be printed in the R console)}

\item{model.lik}{number of classes for thermodynamic integration (see details)}

\item{root.station}{a boolean indicating whether the estimation of the ancestral state should be dropped from the model}

\item{scaleHeight}{A boolean indicating if the trees should be scaled to have a total length of 1}
}
\value{
An object of class jive.
}
\description{
This function makes a jive object from a matrix of intraspecific observations and species phylogeny. The obtained jive object can than be used as an input to \code{\link{jiveMCMC}} function. Intraspecific observations should be stored as matrix, where lines are vector of observations for each species,
with NA for no data. Phylogenetic tree can be either a simmap object (\code{\link{make.simmap}}) or phylo object (\code{\link{as.phylo}}).
}
\details{
This function creates a jive object needed for \code{\link{jiveMCMC}} function.
Trait values must be stored as a matrix, where lines are vectors of observations for each species, with NA for no data.
Rownames are species names. Phylogenetic tree must be provided as either simmap object (for models with multiple regimes)
or as a phylo object for models with one regimes (WN, BM1 or OU1 models). Rownames and tip labels of a phylogenetic tree should match exactly.

There are six models implemeted for the evolution of species means and species variances. The white noise model assumes no phylogenetic correlation
in the evolution of means of variance; models with one selective regime (WN) or multiple regimes (WNM) are available. The models BM1 and BMM are 
implement the Brownian motion model with respectivement one and multiple selective regimes. For more information on the fitting of BMM models
with and without the root.station option set to true, users should refer to the \code{\link{OUwie}} function of the OUwie package. Finally, the models OU1
and OUM implement the Ornstein-Ouhlenbeck models of evolution with respectively one and multiple regimes. Species-specific distribution are modeled as multivariate normal distribution.

Because the ancestral states (at the root of the phylogeny) are difficult to estimate accurately and can someties results in other parameters being 
estimated strangely, we remcommend to drop them from the models (root.station=TRUE), which is the default. For more information on this parameter, check the 
\code{\link{OUwie}} function.
}

\author{Anna Kostikova and Simon Joly}

\examples{
library(OUwie)
library(phytools)
library(MASS)
## number of species we want to simulate
n <- 50

## generate tree with a pure birth model and scale it to the height of 1
tree  <- pbtree(b = 1, n = n, scale = 1, nsim = 1, ape = TRUE)
treeb <- tree

## set parameters for OU1 model of species-specific variances
sig.sq <- 0.9
alpha  <- 0.1
theta0 <- 1
theta  <- 5

## set parameters for BM model of specific-specific means
sig.sq.bm <- 0.5
mu0       <- 350

## set mean number of observations per species
mean.obs <- 20

## get selective regimes (all 1s because of OU1 model)
y <- data.frame(tree$tip.label, rep(1, n))

## add node labels
tree$node.label <- rep("1", n-1)

## simulate species-specific variances
sigma.val <- abs(OUwie.sim(tree, y, simmap.tree=FALSE,
scaleHeight=TRUE, alpha=rep(alpha,2),
sigma.sq=rep(sig.sq,2), theta0=theta0, theta=theta)$X)

## simulate species-specific means
mean.val <- mvrnorm(mu=rep(mu0, length(tree$tip)), Sigma=(sig.sq.bm * vcv(tree)))

## draw a random number of intraspecific observations for each species
spec.obs <- rpois(n, mean.obs)

## generate a data matrix where rows are species and columns are individual observations
traits <- matrix(rnorm(max(spec.obs) * n, mean=mean.val, sd=sqrt(sigma.val)),
nrow=n, ncol=max(spec.obs))
traits <- cbind(as.matrix(max(spec.obs) - spec.obs), traits)

## function to replace empty cells with NA
foo <- function(x){
		to <- x[1]
		x[1:(to + 1)] <- NA
		return(x[-1])
}

## apply to data matrix
traits <- as.matrix(t(apply(traits, 1, foo)))

## add species names to rownames
rownames(traits) <- tree$tip.label
my.jive <- jiveMake(treeb, traits,  model.var="OU1", model.mean="BM", model.lik="Multinorm")
}
