% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcondx.R
\name{lcondx}
\alias{lcondx}
\title{Compute Negative Log-Likelihood}
\usage{
lcondx(jtrt, epsy, mu_xst, start_x, nx, start_y, sgmaj, rhoj)
}
\arguments{
\item{jtrt}{treatment j}

\item{epsy}{residual vector of y \eqn{\epsilon = y - \mu_y}}

\item{mu_xst}{mean vector of x*}

\item{start_x}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in data vector}

\item{nx}{number of observations}

\item{start_y}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in outcome vector}

\item{sgmaj}{variance under treatment j}

\item{rhoj}{correlation coefficient under treatment j}
}
\value{
the log-Likelihood value for given point
}
\description{
Internal. Computes normal, conditional negative Loglikelihood for all x observations with treatment j
}
\seealso{
Other loglikelihoods: \code{\link{llikj}},
  \code{\link{lmvnorm}}, \code{\link{lnorm}},
  \code{\link{negllik_jt_rho}}, \code{\link{negllik_jt}}
}
\concept{loglikelihoods}
