% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikj.R
\name{llikj}
\alias{llikj}
\title{Compute log-Likelihood}
\usage{
llikj(sgmaj, rhoj, j, epsy, mu_xst, start_x, nx, start_y)
}
\arguments{
\item{sgmaj, rhoj}{vector of \eqn{2*Tmax} parameter of the variance/covariance matrix}

\item{j}{treatment}

\item{epsy}{residuals on y}

\item{mu_xst}{mean vector of x*}

\item{start_x}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in data vector}

\item{nx}{number of observations}

\item{start_y}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in outcome vector}
}
\value{
The log-likelihood value corresponding to input values.
}
\description{
Internal. Computes the negativ log-likelihood for all observations with treatment j
}
\seealso{
Other loglikelihoods: \code{\link{lcondx}},
  \code{\link{lmvnorm}}, \code{\link{lnorm}},
  \code{\link{negllik_jt_rho}}, \code{\link{negllik_jt}}
}
\concept{loglikelihoods}
