% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmvnorm.R
\name{lmvnorm}
\alias{lmvnorm}
\title{Compute log-Likelihood of multivariate Normals}
\usage{
lmvnorm(y, mu, Sigmainv)
}
\arguments{
\item{y}{outcome vector}

\item{mu}{vector of means}

\item{Sigmainv}{inverse of variance matrix}
}
\value{
Log-Likelihood value under multivariate Normal assumption.
}
\description{
Internal.
}
\seealso{
Other loglikelihoods: \code{\link{lcondx}},
  \code{\link{llikj}}, \code{\link{lnorm}},
  \code{\link{negllik_jt_rho}}, \code{\link{negllik_jt}}
}
\concept{loglikelihoods}
