% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm.R
\name{lnorm}
\alias{lnorm}
\title{Compute Normal log-Likelihood}
\usage{
lnorm(x, mu, varinv)
}
\arguments{
\item{x}{sample}

\item{mu, varinv}{mean, variance parameters}
}
\value{
The negative log-Likelihood of given parameters.
}
\description{
\code{lnorm} computes log-Likelihood of a Normal distributed sample
}
\seealso{
Other loglikelihoods: \code{\link{lcondx}},
  \code{\link{llikj}}, \code{\link{lmvnorm}},
  \code{\link{loglik_mvnorm_balpan}},
  \code{\link{negllik_jt_rho}}, \code{\link{negllik_jt}}
}
