% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_mvnorm_balpan.R
\name{loglik_mvnorm_balpan}
\alias{loglik_mvnorm_balpan}
\title{Compute log-Likelihood of Balanced Panels}
\usage{
loglik_mvnorm_balpan(y, mu = matrix(0, dim(y)[1], dim(y)[2]), Sigmainv)
}
\arguments{
\item{y}{data}

\item{mu}{mean matrix}

\item{Sigmainv}{inverse variance matrix}
}
\value{
A scalar value, the log-Likelihood for given data, mu and Sigma.
}
\description{
\code{loglik_mvnorm_balpan} computes for balanced panel data
\eqn{dim(y) = t x N} with equal \eqn{\Sigma} across observations
}
\seealso{
Other loglikelihoods: \code{\link{lcondx}},
  \code{\link{llikj}}, \code{\link{lmvnorm}},
  \code{\link{lnorm}}, \code{\link{negllik_jt_rho}},
  \code{\link{negllik_jt}}
}
