% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhStep.R
\name{mhStep}
\alias{mhStep}
\title{Computes a Metropolis Hastings Step}
\usage{
mhStep(theta_old, theta_new, lpost_old, lpost_new, lq_new, lq_old)
}
\arguments{
\item{theta_old}{old proposal parameter value}

\item{theta_new}{new proposal parameter value}

\item{lpost_old}{old log posterior}

\item{lpost_new}{new log posterior}

\item{lq_new}{log q(theta_new|theta_old) the proposal log-densities}

\item{lq_old}{log q(theta_old|theta_new) the proposal log-densities}
}
\value{
A list of the new theta accepted value, or the old
value if rejected and an acceptance indicator.
}
\description{
Internal. Performs Metropolis Hastings steps for
internal MCMC procedure. lpost_old/lpost_new are log-posteriors at
theta_old/theta_new
}
