% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negllik_jt.R
\name{negllik_jt}
\alias{negllik_jt}
\title{Compute Negative log-Likelihood}
\usage{
negllik_jt(theta, j, t, epsy, mu_xst, start_x, nx, start_y, sgmaj, rhoj)
}
\arguments{
\item{theta}{parameter of variance/covariance matrix of dimension 2Tmax x 1}

\item{j}{treatment}

\item{t}{panel time}

\item{epsy}{residuals of y}

\item{mu_xst}{mean vector of x*}

\item{start_x}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in data vector}

\item{nx}{dimension of x}

\item{start_y}{Tmax x trt matrix, containing indicators for
panel time - treatment combinations in outcome vector}

\item{sgmaj}{variance for treatment j}

\item{rhoj}{correlation for treatment j}
}
\value{
A scalar which is the negative log-Likelihood at the given point.
}
\description{
Internal.
}
\seealso{
Other loglikelihoods: \code{\link{lcondx}},
  \code{\link{llikj}}, \code{\link{lmvnorm}},
  \code{\link{lnorm}}, \code{\link{negllik_jt_rho}}
}
\concept{loglikelihoods}
