% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hp.R
\name{plot_hp}
\alias{plot_hp}
\title{plot Hyper-prior function}
\usage{
plot_hp(hpf = "Uniform", hp.pars = c(1, 2), ...)
}
\arguments{
\item{hpf}{name of a density function. Supported density functions are: Uniform, Gamma and Normal}

\item{hp.pars}{a vector of density function parameters}

\item{...}{additional parameters that can be passed to a density function and  \code{\link{par}}}
}
\value{
plot
}
\description{
This function plots a hyper-prior density function. 
Currently supported density function are Uniform, Gamma, Normal, Loggamma and Lognormal. 
The resulting function is used during MCMC \code{\link{mcmc_jive}}
to estimate parameters of priors.
}
\details{
There are three currently implemented density function: 
Uniform, Gamma and Normal. Each of these densities requires two input parameters and hp.pars 
must be a vector of two values and cannot be left empty.
}
\examples{
my.hp <- plot_hp(hpf="Uniform", hp.pars=c(1,2))
}
\author{
Theo Gaboriau
}
