% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jive.R
\name{plot_jive}
\alias{plot_jive}
\title{plot input data from a jive object}
\usage{
plot_jive(jive, col.map = NULL, col = "lightgrey",
  show.tip.label = T, direction = "rightwards", trait.lab = "",
  trait.lim = NULL, srt.label = 0, tip.color, ...)
}
\arguments{
\item{jive}{a jive object built with the function \code{\link{make_jive}}}

\item{col.map}{a vector of mode character indicating colors of the edges for the map ploting. It should be of same size than the number of regimes}

\item{col}{a character indicating the color of the vioplots}

\item{show.tip.label}{a logical indicating whether to show the tip labels on the phylogeny (defaults to TRUE, i.e. the labels are shown).}

\item{direction}{a character string specifying the direction of the tree. Two values are possible: "rightwards" (the default) and "upwards".}

\item{trait.lab}{a charachter specifying the axis label for the traits}

\item{trait.lim}{a vector of mode numeric indicating the limits for trait ploting}

\item{srt.label}{an integer indicating the string rotation in degrees for the tip labels}

\item{tip.color}{the colours used for the tip labels, eventually recycled.}

\item{...}{additional parameters that can be passed to \code{\link{vioplot}}}
}
\value{
plot
}
\description{
This function plots the phylogenetic tree, the trait data and the map used as an input for a jive analysis
}
\examples{
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

my.jive <- make_jive(Anolis_tree, Anolis_traits,  model.var="OU", model.mean="BM")
par(cex.lab = .8, cex.axis = .8, las = 1, mgp = c(2,0.5,0))
plot_jive(jive = my.jive, show.tip.label = T, trait.lab = "Snout to vent length (cm)", srt.label = 0)
my.jive <- make_jive(Anolis_tree, Anolis_traits, Anolis_map*Anolis_tree$edge.length, model.var="OUM", model.mean="BM")
par(cex.lab = .8, cex.axis = .8, las = 1, mgp = c(2,0.5,0))
plot_jive(jive = my.jive, show.tip.label = T, trait.lab = "Snout to vent length (cm)", srt.label = 70, direction = "upwards")
}
\author{
Theo Gaboriau
}
