% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc_jive.R
\name{plot_mcmc_jive}
\alias{plot_mcmc_jive}
\title{Plot \code{\link{jive_mcmc}} results}
\usage{
plot_mcmc_jive(mcmc.log, type = c("trace", "density"), burnin = NULL,
  variable = "log.lik", label = "log(Likelihood)", col = "#000000",
  ...)
}
\arguments{
\item{mcmc.log}{Any mcmc sample with the saved iterations in rows and the variables in columns}

\item{type}{character taken in c("trace", "density"). If both are specified, they are plotted side by side in the same graphical device}

\item{burnin}{The size of the burnin in number of iterations}

\item{label}{Full variable name to be plotted}

\item{...}{Other graphical parameters to parse to \code{\link{par()}}}

\item{var}{The name or number of the variable to plot}
}
\description{
This function plots the trace and/or density of each mcmc sample
}
\details{

}
\examples{

## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

## Run a simple MCMC chain
my.jive <- make_jive(Anolis_tree, Anolis_traits,  model.var="OU", model.mean="BM", root.station = T)
mcmc_jive(my.jive, log.file="my.jive_MCMC.log", sampling.freq=10, print.freq=100, ngen=50000) 

## import the results in R
logfile <- "my.jive_MCMC.log"
res <- read.csv(logfile, header = T, sep = "\\t")

## plot the results
for(variable in colnames(res)[-c(1, 2, 4, 5, 44, 45)]){
 plot_mcmc_jive(res, burnin = 0, variable = variable, lab = variable)
}
}
\author{
Theo Gaboriau
}
