% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmom_help_reg.R
\name{postmom_help_reg}
\alias{postmom_help_reg}
\title{postmom_help_reg method}
\usage{
postmom_help_reg(x, y, Wx, Wy, start_x, nx, start_y, sgma, rho, lambda, D)
}
\arguments{
\item{x}{data}

\item{y}{outcome}

\item{Wx}{feature matrix for selection model}

\item{Wy}{feature matrix for outcome data model}

\item{start_x}{matrix containing indices where certain observations
begin in the vector sorted by treatment and panel times}

\item{nx}{dimension of x}

\item{start_y}{matrix containing indices where certain outcomes
begin in the vector sorted by treatment and panel times}

\item{sgma}{variance parameter matrix}

\item{rho}{correlation parameter matrix}

\item{lambda}{matrix of factor loadings}

\item{D}{initial inclusion for treatment j}
}
\value{
List of Matrices used for computation of posterior moments.
}
\description{
Computes help quantities for posterior moments of regression coefficients.
}
