% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_unbal.R
\name{read_data_unbal}
\alias{read_data_unbal}
\title{Read Data file and transform}
\usage{
read_data_unbal(infile, type, covars = list(NULL), ...)
}
\arguments{
\item{infile}{a list of base matrix and panel matrix}

\item{type}{contains a string of possible model types ("SF")}

\item{covars}{contains additional information about covariate names (optional)}

\item{...}{control parameter passed down in funcional hierarchy}
}
\value{
A data object for further computation.
}
\description{
\code{read_data_unbal} takes object of base matrix and full panel matrix
to read, structure and organise the data into a useable data.frame R object
}
