% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_trt_switchreg.R
\name{select_trt_switchreg}
\alias{select_trt_switchreg}
\title{Run treatment computations with switching regression model}
\usage{
select_trt_switchreg(data, model, prior, mcmc)
}
\description{
Run treatment computations with switching regression model
}
