% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTreatmentSWR.R
\name{selectTreatmentSWR}
\alias{selectTreatmentSWR}
\title{Run treatment computations with switching regression model}
\usage{
selectTreatmentSWR(data, model, prior, mcmc)
}
\description{
Internal.
}
