% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_trt_unbal.R
\name{start_trt_unbal}
\alias{start_trt_unbal}
\title{Estimation of Coefficients and Effects}
\usage{
start_trt_unbal(infile, type, covars)
}
\arguments{
\item{infile}{a data.frame file. Needs to be already correctly structured.}

\item{type}{contains one of currently one possible Strings
("SF"-Shared Factor Model)}

\item{covars}{contains a character vector, naming the covariates of the dataset.

The function takes 2 main parameters, infile containing the data set on
which to perform estimation. The dataset has to have certain properties.
See ?read_data_unbal for more information on how to prepare the dataset
accordingly. For according datasets, the function first sets a lot of parameters
influencing estimation and MCMC sampling process. For the time being this
package includes the functionality of the Shared Factor Model.
One may also be interested in \code{\link{select_trt_sharedfac}} to see how the
function operates.}
}
\value{
For an adequate input file, will output a list object containing all
coefficients of the model.
}
\description{
\code{start_trt_unbal} performs an estimation of treatment effects on panel
structured data. The dataset may be unbalanced, but needs to fullfill
certain conditions to be properly useable.  The function offers
some model choices, which provide the framework in which the treatment effect
is to be estimated.
}
