% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_vals.R
\name{start_vals}
\alias{start_vals}
\title{Computes starting values for parameter search}
\usage{
start_vals(model, x, y, Tn)
}
\arguments{
\item{model}{Contains model parameters}

\item{x}{contains data}

\item{y}{contains outcome (target) information}

\item{Tn}{panel times dimension}
}
\value{
Returns a list of
        \enumerate{
        \item starting values for \eqn{\alpha_\nu}
        \item starting values for \eqn{\beta}
        \item residual variance
}
}
\description{
Computes starting values for alpha and beta parameters. Alpha parameters
are estimated by a GLM fit, beta parameters by a linear model.
}
