% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_jive.R
\encoding{UTF-8}
\name{xml_jive}
\alias{xml_jive}
\title{Write xml file with jive}
\usage{
xml_jive(jive, xml, out = sprintf("\%s_edited.xml", gsub(".xml", "",
  xml)))
}
\arguments{
\item{jive}{an object of class "jive" (see details)}

\item{xml}{name of the output file that will store the log of MCMC chain}

\item{out}{where to write the edited xml}
}
\value{
none
}
\description{
Modifies a xml file from beauti to include a jive model in the Beast 2 analysis
}
\details{
This function takes a xml file generated with Beauti and a jive object generated with \code{\link{make_jive}}
Only jive objects that use models supported by the Beast implementation of jive ("BM", c("BM", "sigma"), "WN", "OU", c("OU", "theta"), c("OU", "root"), c("OU", "root", "theta"))
}
\author{
Theo Gaboriau
}
