% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_jive.R
\name{control_jive}
\alias{control_jive}
\title{Create a list that can be used to tune jive mcmc algorithm}
\usage{
control_jive(level = c("lik", "prior.mean", "prior.var"),
  model.evo = c("BM", "OU", "WN", "OUM", "BMM", "WNM"), traits, map,
  window.size = NULL, initial.values = NULL, proposals = NULL,
  hyperprior = NULL, root.station = F)
}
\arguments{
\item{level}{character taken in c("lik", "prior.mean", "prior.var") to specify on which level of the jive model, the control will operate (see details)}

\item{model.evo}{character taken in c("OU", "BM", "WN", "OUM", "BMM", "WNM") specifying the evolutionary model. ignored if level == "lik"}

\item{traits}{matrix of traits value (see details)}

\item{map}{matrix mapping regimes on every edge of phy}

\item{window.size}{initial window size for proposals during the mcmc algorithm. matrix or vector depending on the value of level and nreg (see details)}

\item{initial.values}{starting parameter values of the mcmc algorithm. matrix or vector depending on the value of level and nreg (see details)}

\item{proposals}{vector of characters taken in c("slidingWin", "slidingWinAbs", "logSlidingWinAbs","multiplierProposal", "multiplierProposalLakner","logNormal", "absNormal") to control proposal methods during mcmc algorithm (see details)}

\item{hyperprior}{list of hyperprior functions that can be generated with \code{\link{hpfun}}function. Ignored if level == "lik" (see details)}

\item{root.station}{boolean indicating whether the theta_0 should be dropped from the OU or OUM models}
}
\value{
A list to parse into control argument of \code{\link{make_jive}} function. The list is containing the following objects:
$model : a function to calculate the likelihood ($lik) or the priors ($prior.mean, $prior_var)
$ws : a list containing the window size of proposals for each estimated parameter
$init : a list containing the starting value of the mcmc chain for each parameter
$prop : a list containing the proposal functions for the update of each parameter
$map (only if level %in% c(prior.mean, prior.var) : a matrix containing the mapping of regimes onto branch for a specific model
$hprior (only if level %in% c(prior.mean, prior.var) : a list containing the hyperprior function for each parameter
}
\description{
This function creates an object to parse in the control argument of the \code{\link{make_jive}} function. The output will be different regarding which level of the jive model the user wants to tune ($lik, $prior.mean, $prior.var). This function allows tuning of : initial window size for proposals, starting parameter value, proposal methods and Hyperprior specification
}
\details{
If arguments initial.ws, initial.pv, proposals or hyperprior are left blank, the default tuning that we found appropriate for most datasets is applied

If level == "lik"
initial.ws and initial.pv must be entered as a matrix with 2 columns (respectively mean and variance) and a number of rows equal to the number of species. proposal must be a vector of size 2 (respectively mean and variance)

If level == "prior.mean" or "prior.var"
initial.ws and initial.pv must be entered as a vector of variable size depending on the chosen evolutionary model. for OU and OUM, the window size and parameter values must be entered in the following order c(alpha, sigma, theta0, theta1, ..., thetaN). for BM, BMM, WN and WNM, the window size and parameter values must be entered in the following order c(sigma1, ..., sigmaN, theta0), proposal must be a vector of size three for OU and OUM c(alpha, sigma, thetas) and of size two for BM, BMM, WN and WNM c(sigmas, theta)

proposals
Has to be one the following : "slidingWin" for Sliding window proposal unconstrained at maximum, "multiplierProposal", for multiplier proposal

Hyperprior
list of hyperpriror functions (see \code{\link{hpfun}}). User must provide a list of size 2 for BM, BMM, WN and WNM (sigmas, theta0) and of size 3 for OU and OUM (alpha, sigma, thetas)
}
\examples{

}
\author{
Theo Gaboriau
}
