% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_loadings.R
\name{draw_loadings}
\alias{draw_loadings}
\title{Helpfunction for Model SRF latent Factor Switching Regression}
\usage{
draw_loadings(resy, Xf, resx, start_x, nx, start_y, sgma, rho, prior)
}
\description{
Internal.
}
