% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_jive_traits.R
\name{format_jive_traits}
\alias{format_jive_traits}
\title{Format traits matrix to parse into \code{\link{make_jive}} function}
\usage{
format_jive_traits(obs)
}
\arguments{
\item{obs}{a dataframe with species names in the first column and individual trait values in the second columns}
}
\description{
This function takes a dataframe with indivdual observations ans species names and returns a traits matrix to to parse into \code{\link{make_jive}} function
}
\details{

}
\examples{

}
\author{
Theo Gaboriau
}
