% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jive.R
\encoding{UTF-8}
\name{make_jive}
\alias{make_jive}
\title{Create a list that can be used as an input to mcmc_bite}
\usage{
make_jive(phy = NULL, traits, map = NULL, model.mean = c("BM"),
  model.var = c("OU"), scale = F, control = list(), nreg = NULL)
}
\arguments{
\item{phy}{phylogenetic tree provided as either a simmap or a phylo object}

\item{traits}{matrix of traits value for every species of phy (see details)}

\item{map}{matrix mapping regimes on every edge of phy (see details)}

\item{model.mean}{model specification for trait mean evolution. Supported models are "OU", "BM", "WN". The user can also specify if the assumptions of the model should be relaxed (see details)}

\item{model.var}{model specification for trait variance evolution. Supported models are "OU", "BM", "WN". The user can also specify if the assumptions of the model should be relaxed (see details)}

\item{scale}{boolean indicating whether the tree should be scaled to unit length for the model fitting}

\item{control}{list to control tuning parameters of the MCMC algorithm (see details)}

\item{nreg}{integer giving the number of regimes for a Beast analysis. Only evaluated if phy == NULL}
}
\value{
A list of functions and tuning parameters to parse into \code{\link{mcmc_bite}} function.
}
\description{
This function creates a jive object from a matrix of intraspecific observations
and species phylogeny. The obtained jive object is a list that can than be used as an input to \code{\link{mcmc_bite}} function
Intraspecific observations should be stored as matrix, where lines are vector of observations for each species,
with NA for no data. Phylogenetic tree can be either a simmap object (\code{\link{make.simmap}}) or phylo object (\code{\link{as.phylo}})
}
\details{
This function creates a jive object needed for \code{\link{mcmc_bite}} function.  
Trait values must be stored as a matrix, where lines are vectors of observations for each species, with NA for no data. Rownames are species names that should match exactly tip labels of the phylogenetic tree.

Phylogenetic tree must be provided as either simmap object or as a phylo object. If the phylogenetic tree is a phylo object but model specification indicates multiple regimes, user must provide a mapping of the regime in map. If you keep the phy = NULL options the JIVE object can only be parsed to the \code{\link{xml_bite}} function.

map is a matrix giving the mapping of regimes on phy edges. Each row correspond to an edge in phy and each column correspond to a regime. If map is provided the map from the simmap object is ignored.   

variance and mean evolution can be modeled with Ornstein-Uhlenbeck (OU), Brownian Motion (BM) or White Noise (WN) processes. Multiple regimes can be defined for both models and will apply on thetas: model.mean/var = c("OU", "theta"), sigmas: model.mean/var = c("OU", "sigma") or stationary variances: model.mean/var = c("OU", "sv") for OU and on sigmas only for WN: model.mean/var = c("WN", "sigma") and BM: model.mean/var = c("BM", "sigma"). While using the OU model, the user can also relax the stationarity of the root: model.mean/var = c("OU", "root") and relax several assumptions at the same tme model.mean/var = c("OU", "root", "theta") 
Species-specific distributions are modeled as multivariate normal distributions

control is a list containig tuning parameters acting at different levels of the MCMC algorithm ($lik for likelihood level, $prior.mean for mean prior level and $prior.var for variance prior level). Inside each level ($lik, $prior.mean, $prior.var), the user can modify the default value of initial parameter value ($pv), initial window size ($ws), proposal methods ($prop) for $lik, $prior.mean and $prior.var and hyperpriors ($hprior) for $prior.mean and $prior.var. 
The \code{\link{control_jive}} function provides an easy way to modify control parameters (see examples) 

parameters used in the different models:

White Noise model (WN):
\itemize{
 \item theta0: root value, abbreviated wn.the
 \item sigma square: evolutionary rate, abbreviated wn.sig or wn.sig.1, wn.sig.2, ..., wn.sig.n for n regimes if "sigma" is specified in model.mean/var
}
 
Brownian Motion model (BM):
\itemize{
 \item theta0: root value, abbreviated bm.the
 \item sigma square: evolutionary rate, abbreviated bm.sig or bm.sig.1, bm.sig.2, ..., bm.sig.n for n regimes if "sigma" is specified in model.mean/var
}

Ornstein Uhlenbeck model (OU):
\itemize{
 \item theta0: root value, abbreviated ou.the.0. Only used if "root" is specified in model.mean/var
 \item sigma square: evolutionary rate, abbreviated ou.sig or ou.sig.1, ou.sig.2, ..., ou.sig.n for n regimes if "sigma" is specified in model.mean/var
 \item optimal value, abbreviated ou.the.1 or ou.the.1, ou.the.2, ..., ou.the.n for n regimes if "theta" is specified in model.mean/var
 \item stationary variance (alpha/2*sigma_sq with alpha being the strength of selection), abbreviated ou.sv or ou.sv.1
}
}
\examples{

## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

## JIVE object to run jive with single regimes
my.jive <- make_jive(phy = Anolis_tree, traits = Anolis_traits,
 model.mean="BM", model.var= c("OU", "root"))

## JIVE object to run jive with multiple regimes
my.jive <- make_jive(Anolis_tree, Anolis_traits, map = Anolis_map,
 model.mean="BM", model.var=c("OU", "theta", "sv"))

## JIVE object to run jive from an ancestral state reconstruction (stochastic mapping)
# First generate simmap object
library(phytools)
n= length(Anolis_tree$tip.label)
trait = rep(0,n)
trait[c(4,3,14,16, 6,5)] = 1
names(trait) =  Anolis_tree$tip.label

mapped_tree=make.simmap(Anolis_tree, trait, model='SYM')
plotSimmap(mapped_tree)

my.jive <- make_jive(mapped_tree, Anolis_traits, model.mean=c("OU"), model.var=c("OU", "theta"))
 
}
\seealso{
\code{\link{xml_bite}}, \code{\link{mcmc_bite}}
}
\author{
Theo Gaboriau, Anna Kostikova, Daniele Silvestro and Simon Joly
}
