% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regmat_unbal.R
\name{make_regmat_unbal}
\alias{make_regmat_unbal}
\title{Transscribes Basic Object to Internal Matrix}
\usage{
make_regmat_unbal(data, model)
}
\arguments{
\item{data}{The collected information and data in one internal object}

\item{model}{Information on the model type, etc.}
}
\value{
A modelling object with W, the regressor matrix of dimension n*t+1 x p
}
\description{
Internal. Takes the basic objects after read-in of base.mat
and panel.mat and restructures them to an internal data object to proceed.
Constructs a proper model matrix, which encapsulates the outcome model
covariates for each panel time
}
