% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_lik.R
\name{marginal_lik}
\alias{marginal_lik}
\title{Calculate marginal likelihood by thermodynamic integration (LTI)}
\usage{
marginal_lik(mcmc.log, burnin = 0)
}
\arguments{
\item{mcmc.log}{the output file of a \code{\link{jive_mcmc}} run}
}
\description{
Calculate the marginal likelihood from a logfile generated by \code{\link{jive_mcmc}} with thermodynamic integration
}
\examples{
## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

## Run a MCMC chain with thermodynamic Integration
set.seed(300)
my.jive <- make_jive(Anolis_tree, Anolis_traits,  model.var="OU", model.mean="BM")
mcmc_jive(my.jive, log.file="my.jive_MCMC.log", ncat=10, sampling.freq=10, print.freq=100, ngen=5000, burnin=500)
 
## import the results in R
logfile <- "my.jive_MCMC.log"
res <- read.csv(logfile, header = T, sep = "\\t")
 
mlik <- marginal_lik(res)
mlik
}
\author{
Theo Gaboriau and Simon Joly
}
