% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bite.R
\encoding{UTF-8}
\name{mcmc_bite}
\alias{mcmc_bite}
\title{MCMC algorithm}
\usage{
mcmc_bite(model, log.file = "bite_mcmc.log", sampling.freq = 1000,
  print.freq = 1000, ncat = 1, beta.param = 0.3, ngen = 5e+06,
  burnin = 0)
}
\arguments{
\item{model}{an object of class "jive" or other objects from the bite package (see details)}

\item{log.file}{name of the output file that will store the log of MCMC chain}

\item{sampling.freq}{sampling frequency of the MCMC chain (how often chain will be saved into output file}

\item{print.freq}{printing frequency of the MCMC chain (how often chain will be printed in the R console)}

\item{ncat}{number of classes for thermodynamic integration (see details)}

\item{beta.param}{beta value to define classes for thermodynamic integration (see details)}

\item{ngen}{number of generation in MCMC chain}

\item{burnin}{a burning phase of MCMC chain (has to be specified for thermodynamic integration)}
}
\value{
none
}
\description{
Implements Markov chain Monte Carlo sampling for trait evolution models
}
\details{
This function runs MCMC sampling on jive object \code{\link{make_jive}} or objects describing other models of the bite package.
The jive object contains both the dataset and set of model to be used in MCMC. This function implements both a conventional MCMC
and an MCMC with thermodynamic integration. The latter option is turned off by default and can be changed by
setting ncat to values > 1. The recommended ncat for TI is 10. When setting ncat > 1, make sure to specify burning.
As a rule of thumb set burning to 1/10 fraction of ngen.
}
\examples{
\dontrun{
 ## Load test data
 data(Anolis_traits)
 data(Anolis_tree)
 data(Anolis_map)
 
 ## Run a simple MCMC chain
 set.seed(300)
 my.jive <- make_jive(Anolis_tree, Anolis_traits, map = Anolis_map,
 model.var=c("OU", "root", "theta", "alpha"), model.mean="BM")
 mcmc_bite(my.jive, log.file="my.jive_MCMC.log",
 sampling.freq=10, print.freq=100, ngen=5000) 

 my.jive <- make_jive(Anolis_tree, Anolis_traits, map = Anolis_map,
 model.var=c("OU", "root", "theta"), model.mean="BM")
 mcmc_bite(my.jive, log.file="my.jive_MCMC.log",
  sampling.freq=10, print.freq=100, ngen=5000) 
  
 ## Run an MCMC chain with thermodynamic integration
 mcmc_bite(my.jive, log.file="my.jive_MCMC_TI.log", ncat=10, 
  sampling.freq=10, print.freq=100, ngen=5000, burnin=500) 
}
}
\author{
Theo Gaboriau, Anna Kostikova, Daniele Silvestro, and Simon Joly
}
