% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bf.R
\name{plot_bf}
\alias{plot_bf}
\title{Plots summary of Bayes Factors calculations}
\usage{
plot_bf(m.liks, thr = 2, dir = c("vertical", "horizontal"),
  col = c("#000000", "#2e86ab", "#d32f23"), col.thr = c("#a6e1fa"),
  ax.lab = "Bayes Factor", main = "", mod.lab = "default",
  rank = T, group = c("mean", "var"), cex = c(1, 1, 2),
  lwd.l = c(1, 1, 2), srt.lab = 0, adj.lab = 0,
  col.axis = c("#000000", "#2e86ab", "#d32f23"), cex.axis = c(0.8, 0.8,
  1), space = c(1.2, 0.8))
}
\arguments{
\item{m.liks}{matrix of marginal likelihoods (see details)}

\item{thr}{value of BF threshold for model selction (default is 2)}

\item{col}{color of the lines and dots. Could be of size one two or three. The first element applies to models that are above thr, the last element applies to the best model, the middle element applies to models that are below thr}

\item{col.thr}{color for the threshold area}

\item{rank}{should models be ranked by BF scores? (default is TRUE)}

\item{group}{only evaluated if rank == F, character indicating whether the models should be grouped regarding the evolutionnary model on "mean" or "var"}

\item{cex}{size of the dots. Could be of size one two or three. The first element applies to models that are above thr, the last element applies to the best model, the middle element applies to models that are below thr}

\item{lwd.l}{thickness of the lines. Could be of size one two or three. The first element applies to models that are above thr, the last element applies to the best model, the middle element applies to models that are below thr}

\item{col.axis}{color of axis annotation. Could be of size one two or three. The first element applies to models that are above thr, the last element applies to the best model, the middle element applies to models that are below thr}

\item{space}{only evaluated if rank == F, numeric(2) indicating the space between groups and the space between members of the same groups}

\item{ylab}{name of the models. If ylab = "default" the dimnames of m.liks are used}
}
\description{
Lolipop plot representing values of BF (Bayes Factor) scores for different models
}
\examples{
## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

## Run a MCMC chain with thermodynamic Integration
mean_models <- c("BM", "OU")
var_models <- c("BM", "OU", "BMM", "OUM")
mliks <- matrix(0, nrow = 2, ncol = 4, dimnames = list(mean_models, var_models))

for(mn in mean_models){
 for(vr in var_models){
  my.jive <- make_jive(Anolis_tree, Anolis_traits,  model.var=vr, model.mean=mn)
  logfile <- sprintf("my.jive_MCMC_M-\%s_V-\%s.log", mn, vr)
  mcmc_jive(my.jive, log.file=logfile, ncat=10, sampling.freq=10, print.freq=100, ngen=5000, burnin=500)
  res <- read.table(logfile, header = T, sep = "\\t")
  mliks[mn,vr] <- marginal.lik(res)
 }
}

plot_bf(mliks)
plot_bf(mliks, thr = 4)
plot_bf(mliks, thr = 4, rank = F, group = "mean")
plot_bf(mliks, dir = "horizontal", srt.lab = -60, adj.lab = c(.3,-1))
}
\author{
Theo Gaboriau
}
