% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc_bite.R
\encoding{UTF-8}
\name{plot_mcmc_bite}
\alias{plot_mcmc_bite}
\title{Plot \code{\link{mcmc_bite}} results}
\usage{
plot_mcmc_bite(mcmc.log, type = c("trace", "density"), burnin = 0,
  variable = "log.lik", label = ifelse(is.numeric(variable),
  colnames(mcmc.log)[variable], variable), col = "#000000",
  cex.est = 1, bty = "n", ...)
}
\arguments{
\item{mcmc.log}{Any mcmc sample with the saved iterations in rows and the variables in columns}

\item{type}{Character taken in c("trace", "density"). If both are specified, they are plotted side by side in the same graphical device}

\item{burnin}{The size of the burnin in number of iterations or the proportion of iteration you want to remove}

\item{variable}{The name or number of the variable to plot}

\item{label}{Full variable name to be plotted}

\item{col, bty, ...}{Other graphical parameters to parse to \code{\link[graphics]{par}}}

\item{cex.est}{The magnification to be used for estimates display}
}
\description{
This function plots the trace and/or density of each mcmc sample
}
\examples{

## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

## Run a simple MCMC chain
my.jive <- make_jive(Anolis_tree, Anolis_traits,  model.var="OU", model.mean="BM")
mcmc_bite(my.jive, log.file="my.jive_mcmc.log", sampling.freq=10, print.freq=100, ngen=50000) 

## import the results in R
logfile <- "my.jive_mcmc.log"
res <- read.csv(logfile, header = TRUE, sep = "\\t")

## plot the results
plot_mcmc_bite(res, burnin = 0.2, variable = "prior.mean", cex.est = .7)

}
\author{
Theo Gaboriau
}
