% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvo.R
\name{plot_pvo}
\alias{plot_pvo}
\title{Plots estimates of species traits distribution}
\usage{
plot_pvo(traits, mcmc.log, tip = 1, label = "tip 1", conf = 0.95,
  stat = "median", trait.lab = "trait", col = NULL, border = NULL,
  legend = F, lolipop = T, ...)
}
\arguments{
\item{traits}{trait data used to perform the jive analysis. This has to be of the same form as the one used in \code{\link{make_jive}}}

\item{mcmc.log}{the output file of a \code{\link{jive_mcmc}} run}

\item{tip}{An integer giving the row corresponding to the species to be plotted}

\item{label}{A character giving the species name}

\item{conf}{A number of [0,1] giving the confidence level desired.}

\item{stat}{A character giving the function to be used to estimate species mean and variance from the posterior distributions. Must be one of be "mean" and "median"}

\item{trait.lab}{a charachter specifying the axis label for the traits}

\item{col}{color of the density filling. Must be of size two for estimates and HPD. If col and border are NULL, two random colors are assigned}

\item{border}{color of denstiy contour. Could be of size one or two. If border = NULL, no borders are plotted}

\item{lolipop}{a logical specifying wether the sample positions should be presented as lolipops}
}
\description{
Density plot representing estimated species trait distributions under a jive model.
This function plots the mean or median density distribution and the HPD distributions assuming that the trait is normally distributed
}
\examples{
## Load test data
data(Anolis_traits)
data(Anolis_tree)
data(Anolis_map)

}
\author{
Theo Gaboriau
}
