% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_trt_sharedfac.R
\name{select_trt_sharedfac}
\alias{select_trt_sharedfac}
\title{Modelling Treatment Effects via Shared Factor Model}
\usage{
select_trt_sharedfac(data, model, prior, mcmc)
}
\arguments{
\item{data}{takes data object from the function start_trt_unbal}

\item{model}{takes model parameters from function start_trt_unbal}

\item{prior}{specified prior parameters}

\item{mcmc}{mcmc parameters specified by start_trt_unbal, potentially subject to change}
}
\value{
Returns a MCMC selection list object containing all estimated coefficients and
other meassures.
}
\description{
[Internal Function] \code{select_trt_sharedfac} implements the shared
factor model. It is the core function of wrapper \code{\link{start_trt_unbal}}
which passes its parameters on, concerning the MCMC process, model and data.
}
