% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_jive.R
\name{sim_jive}
\alias{sim_jive}
\title{Simulate JIVE process}
\usage{
sim_jive(phy, map = NULL, model.var = "OU", model.mean = "BM",
  v.pars = list(the = c(2, 1), sig2 = 0.1, alp = 1), m.pars = list(the
  = 0, sig2 = 0.1), sampling = list(min = 0, max = 7))
}
\arguments{
\item{phy}{Phylogenetic tree}

\item{map}{list containing the mapping of regimes over each edge (see details).}

\item{model.var}{model specification for the simulation of trait variance evolution. Supported models are c("OU", "BM", "WN", "OUM", "BMM", "WNM")}

\item{model.mean}{model specification for the simulation of trait mean evolution. Supported models are c("OU", "BM", "WN", "OUM", "BMM", "WNM")}

\item{v.pars}{parameters used for the simulation of trait variance evolution (see details).}

\item{m.pars}{parameters used for the simulation of trait mean evolution (see details).}
}
\value{
returns the simulated mean and variance of the trait for each tip
}
\description{
Generate random values of trait mean and variance simulated under a JIVE process along a phylogenetic tree
}
\details{
map : the list must be ordered in the same order than phy$edge. Each element represents an edge and contains a vector indicating the time spent under each regime in the branch. The name of the regimes must appear on the map
pars : depends on the chosen model.
BM and WN : list containing the ancestral state (theta0) and the evolutionary rate (sigma^2)
BMM and WNM : list containing the ancestral state (theta0) and the different evolutionary rates corresponding to the regimes (sigma^2i)
OU : list containing the ancestral state and optimum (theta0, theta) the evolutionary rate (sigma^2) and the strentgh of selection (alpha)
OUM : list containing the ancestral state and optima (theta0, thetai) the evolutionary rates (sigma^2i) and the strentghs of selection (alphai). At least one of thetas, sigmas^2 or alphas must have as many values as regimes.
}
\examples{

library(phytools)
phy <- pbtree(n = 50)
Q <- cbind(c(-.002, .002), c(.002, -.002))
phy <- sim.history(phy, Q = Q)
# MBM and VOU
jive_phy <- sim_jive(phy, phy$maps)

# MWNM and VOUM
jive_phy <- sim_jive(phy, phy$maps, "OUM", "WNM",  v.pars = list(the = c(10,5,10), 
sig2 = 0.1, alp = c(0.2, 0.8)), m.pars = list(the = 0, sig2 = c(0.1, 0.5)))
}
\author{
Theo Gaboriau
}
