% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_paneldata.R
\docType{data}
\name{trt_paneldata}
\alias{trt_paneldata}
\title{Example dataset containing the panel information}
\format{A list containing numerous objects:
\describe{
 \item{panel.mat}{Data matrix, but other format than base.mat}
       \item ID: id of subject
       \item panelT: timepoint t of meassured features and dependent variable
       \item y: dependent variable, outcome
       \item V1: variable 1, same as V1 in baseline
       \item V2: variable 2, same as V2 in baseline
       \item t2: dummy for panel time = 2
       \item t3: dummy for panel time = 3
       \item t4: dummy for panel time = 4
}}
\source{
This dataset stems from a simulation process and represents fictive data.
}
\usage{
trt_paneldata
}
\description{
A simulated dataset (from the framework of the shared factor model) to
demonstrate the use of the SR model in package 'bite'. The file contains
5000 subjects, each of which has 4 panel observations on simulated variables.
The second of 2 argument datasets for\code{\link{bayesTrtEffects}}.
}
\keyword{datasets}
